requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.39.0
black==23.3.0
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.1
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
croniter==1.3.15
cycler==0.11.0
datasets==2.12.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
dill==0.3.6
docstring-parser==0.15
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.97.0
ffmpy==0.3.0
fire==0.5.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.5.0
gdown==4.7.1
google-auth==2.19.1
google-auth-oauthlib==1.0.0
gradio==3.34.0
gradio_client==0.2.6
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
jsonargparse==4.22.0
jsonschema==4.17.3
kiwisolver==1.4.4
lightning-cloud==0.5.37
lightning-utilities==0.8.0
linkify-it-py==2.0.2
lit==16.0.5.post0
loralib==0.1.1
Markdown==3.4.3
markdown-it-py==2.2.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.9.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.1
peft @ git+https://github.com/huggingface/peft.git@e536616888d51b453ed354a6f1e243fecb02ea08
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.4
pytz==2023.3
readchar==4.0.5
regex==2023.6.3
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
starsessions==1.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
timm==0.9.2
tokenize-rt==5.1.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
triton==2.0.0
typeshed-client==2.3.0
tzdata==2023.3
uc-micro-py==1.0.2
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.4
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
